Correct the invoice amount_to_pay calculation for multiple payments
A typo causes an iteration through payments to set the paid amount to the value
of the last record only instead of tallying all the payment values, so invoices
with multiple payments do not report correct values. This patch corrects the
typo to make the calculation work as intended.

issue6471
review38351002